requirements.txt
pyserial==3.4
pydantic==1.10.7